Refactor map location filtering
