+++ src/features/organizations/components/ActivistPortalEventMap.tsx
@@ -1,85 +1,31 @@
 import { Box, SxProps } from '@mui/material';
 import { Layer, Map, Source } from '@vis.gl/react-maplibre';
 import { Map as MapType } from 'maplibre-gl';
-import { FC, PropsWithChildren, useEffect, useMemo, useState } from 'react';
+import { FC, PropsWithChildren, useMemo, useState } from 'react';
 
 import notEmpty from 'utils/notEmpty';
 import ZUIMapControls from 'zui/ZUIMapControls';
-import { useAppDispatch, useEnv } from 'core/hooks';
+import { useAppSelector, useEnv } from 'core/hooks';
 import { markerImage } from '../utils/markerImage';
 import { pointsToBounds } from 'utils/mapUtils';
 import { ZetkinEventWithStatus } from 'features/home/types';
 import { Latitude, Longitude } from 'features/areas/types';
-import { filtersUpdated } from '../store';
 import { isLocationInGeoJSONFeatures } from '../../map/utils/locationFiltering';
+import useMapClickFiltering from '../hooks/useMapClickFiltering';
 
 export const ActivistPortalEventMap: FC<
   PropsWithChildren<{
     events: ZetkinEventWithStatus[];
-    geojsonToFilterBy: GeoJSON.Feature[];
     sx?: SxProps;
   }>
-> = ({ children, events, geojsonToFilterBy, sx }) => {
+> = ({ children, events, sx }) => {
   const [map, setMap] = useState<MapType | null>(null);
-  const dispatch = useAppDispatch();
 
-  useEffect(() => {
-    if (map) {
-      map.on('click', 'locationMarkers', (event) => {
-        if (event.features) {
-          const geojsonFeatures = event.features.map((feature) => {
-            const location = JSON.parse(feature?.properties?.location);
-            return {
-              geometry: {
-                coordinates: [location.lat, location.lng],
-                type: 'Point',
-              },
-              properties: {
-                location,
-              },
-              type: 'Feature',
-            };
-          }) as GeoJSON.Feature[];
+  const { geojsonToFilterBy } = useAppSelector(
+    (state) => state.organizations.filters
+  );
 
-          const bounds = pointsToBounds(
-            geojsonFeatures.map((feature) => {
-              if (feature.geometry.type === 'Point') {
-                return [
-                  feature.geometry.coordinates[1] as Longitude,
-                  feature.geometry.coordinates[0] as Latitude,
-                ];
-              } else {
-                return [0 as Longitude, 0 as Latitude];
-              }
-            })
-          );
-
-          if (bounds) {
-            map.fitBounds(bounds, {
-              animate: true,
-              duration: 1200,
-              maxZoom: 16,
-              padding: 20,
-            });
-          }
-
-          dispatch(
-            filtersUpdated({
-              geojsonToFilterBy: geojsonFeatures,
-            })
-          );
-        }
-      });
-
-      map.on('mouseenter', 'locationMarkers', () => {
-        map.getCanvas().style.cursor = 'pointer';
-      });
-
-      map.on('mouseleave', 'locationMarkers', () => {
-        map.getCanvas().style.cursor = '';
-      });
-    }
-  }, [map]);
+  useMapClickFiltering(map);
 
   const env = useEnv();
   const bounds = useMemo(
